pyproject.toml
[tool.poetry]
name = "web-at-agents-v1-0"
version = "0.1.0"
description = ""
authors = ["Aidar Kutdusov <aydar2k@mail.ru>"]
readme = "README.md"
packages = [{include = "web_at_agents_v1"}]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
networkx = "^3.4.2"
sortedcontainers = "^2.4.0"
ipythonblocks = "^1.9.1"
jsonpickle = "^3.4.2"
flask-socketio = "^5.4.1"
eventlet = "^0.37.0"
colorama = "^0.4.6"
at-queue = {git = "https://github.com/grigandal625/AT_QUEUE.git", rev = "master"}
at-solver = {git = "https://github.com/grigandal625/AT_SOLVER.git", rev = "master"}
at-temporal-solver = {git = "https://github.com/grigandal625/AT_TEMPORAL_SOLVER.git", rev = "master"}
at-blackboard = {git = "https://github.com/grigandal625/AT_BLACKBOARD.git", rev = "master"}
uvicorn = "^0.32.0"
asgiref = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
